Refine hibernation hook evidence and settings grouping
