+++ .coderabbit.yaml
@@ -18,7 +18,7 @@ reviews:
         Apply the cmux custom Swift lint rules in `.github/review-bot-rules/` during review. Treat those files as the source of truth. Focus on production Swift changes and ignore test-only scaffolding unless it makes production behavior worse.
     - path: "Sources/**/*.swift"
       instructions: |
-        Apply `.github/review-bot-rules/settings-organization-search.md` during review. For settings UI, navigation, search, and command-palette setting changes, flag stale or overloaded categorization, missing old-name aliases, missing JSON-key keywords, and broken section or setting anchors.
+        Apply `.github/review-bot-rules/settings-organization-search.md` during review. For settings UI, navigation, search, and command-palette setting changes, re-evaluate the affected settings area and flag stale or overloaded categorization, missing old-name aliases, missing JSON-key keywords, missing synonyms for muscle memory, and broken section or setting anchors.
     - path: "**/Package.swift"
       instructions: |
         Apply the cmux custom Swift lint rules in `.github/review-bot-rules/`, especially the concurrency modernization, actor isolation, blocking runtime, file/package boundary, and architectural rethink rules.
@@ -88,7 +88,7 @@ reviews:
       - name: "cmux settings organization"
         mode: error
         instructions: |
-          For settings UI, navigation, search, command-palette setting toggles, or persisted settings-key changes, fail when the diff violates `.github/review-bot-rules/settings-organization-search.md`: stale or overloaded categories, moved settings without old-name and JSON-key aliases, broken search anchors, or too few synonyms for common names and adjacent product vocabulary. Pass for implementation-only changes that do not alter the visible settings surface.
+          For settings UI, navigation, search, command-palette setting toggles, or persisted settings-key changes, fail when the diff violates `.github/review-bot-rules/settings-organization-search.md`: stale or overloaded categories, moved settings without old-name and JSON-key aliases, broken search anchors, or too few synonyms for common names, old section names, command-palette wording, CLI/docs names, likely misspellings, and adjacent product vocabulary. Pass for implementation-only changes that do not alter the visible settings surface.
       - name: "cmux SwiftUI state layout"
         mode: error
         instructions: |
